Exploring routing (specifically ngRx Router) in Angular 2 RC 4.
This was a *failed* experiment. It originally started out as the Component
Router; but, the Component Router couldn't do what I needed to do. Then, I
switched over to the ngRx Router, which *could* do what I needed. But, the
ngRx Router has now been end-of-lifed. So, you can't win.